remove use of non emotion/css stylings from sample components
