+++ apps/newsletters-ui/src/app/DataClientTest.tsx
@@ -1,7 +1,7 @@
 import { newslettersDataClient } from '@newsletters-nx/newsletters-data-client';
 
 export const DataClientTest = () => (
-  <div style={{ background: 'yellow' }}>
+  <div>
     <p>the newslettersDataClient function returns:</p>
     <pre>{newslettersDataClient()}</pre>
   </div>
